[BE][Refactor] CommuController 수정 : Custom S3 client 분리로 의존성을 떨어뜨려 놓음
