+++ server/src/main/java/com/codestates/sebmainproject009/auth/controller/JwtController.java
@@ -30,28 +30,32 @@ public class JwtController {
     @PostMapping("/refresh")
     public ResponseEntity postRefresh(@RequestHeader("refresh") String authorizationHeader){
 
+        try {
+            if(jwtTokenizer.verifyTokenExpiration(authorizationHeader)){
 
-        if(jwtTokenizer.verifyTokenExpiration(authorizationHeader)){
+                Jws<Claims> jwsClaims = jwtTokenizer.getClaimsByToken(authorizationHeader);
+                User foundUser = userService.findUser((String) jwsClaims.getBody().get("sub"));
 
-            Jws<Claims> jwsClaims = jwtTokenizer.getClaimsByToken(authorizationHeader);
-            User foundUser = userService.findUser((String) jwsClaims.getBody().get("sub"));
+                Map<String , Object> claims = new HashMap<>();
+                claims.put("username", foundUser.getEmail());
+                claims.put("roles", foundUser.getRoles());
+                claims.put("userId",foundUser.getUserId());
 
-            Map<String , Object> claims = new HashMap<>();
-            claims.put("username", foundUser.getEmail());
-            claims.put("roles", foundUser.getRoles());
-            claims.put("userId",foundUser.getUserId());
+                String subject = foundUser.getEmail();
+                Date expiration = jwtTokenizer.getTokenExpiration(jwtTokenizer.getAccessTokenExpirationMinutes());
 
-            String subject = foundUser.getEmail();
-            Date expiration = jwtTokenizer.getTokenExpiration(jwtTokenizer.getAccessTokenExpirationMinutes());
+                String base64EncodedSecretKey = jwtTokenizer.encodeBase64SecretKey(jwtTokenizer.getSecretKey());
 
-            String base64EncodedSecretKey = jwtTokenizer.encodeBase64SecretKey(jwtTokenizer.getSecretKey());
+                String accessToken = jwtTokenizer.generateAccessToken(claims, subject, expiration, base64EncodedSecretKey);
 
-            String accessToken = jwtTokenizer.generateAccessToken(claims, subject, expiration, base64EncodedSecretKey);
-
-            return new ResponseEntity("Bearer ".concat(accessToken), HttpStatus.CREATED);
-        }else {
-            return new ResponseEntity("UNAUTHORIZED",HttpStatus.UNAUTHORIZED);
+                return new ResponseEntity("Bearer ".concat(accessToken), HttpStatus.CREATED);
+            }else {
+                return new ResponseEntity("UNAUTHORIZED",HttpStatus.UNAUTHORIZED);
+            }
+        }catch (Exception e){
+            return new ResponseEntity(e.getMessage(),HttpStatus.UNAUTHORIZED);
         }
+
 
 
     }
